refactor: made about components folders name consistent
