+++ public/chunk-map.json
@@ -1,1 +1,1 @@
-{"polyfill":["/polyfill.js","/polyfill.c143712b3b39b2a3434c.hot-update.js"],"commons":["/commons.css","/commons.js","/commons.c143712b3b39b2a3434c.hot-update.js"],"component---cache-dev-404-page-js":["/component---cache-dev-404-page-js.js"],"component---src-pages-404-js":["/component---src-pages-404-js.js"],"component---src-pages-about-js":["/component---src-pages-about-js.js","/component---src-pages-about-js.c143712b3b39b2a3434c.hot-update.js"],"component---src-pages-contact-js":["/component---src-pages-contact-js.js"],"component---src-pages-index-js":["/component---src-pages-index-js.js"],"component---src-pages-work-js":["/component---src-pages-work-js.js"]}
+{"polyfill":["/polyfill.js","/polyfill.f72585a3f4cb0feb1cf0.hot-update.js"],"commons":["/commons.css","/commons.js","/commons.f72585a3f4cb0feb1cf0.hot-update.js"],"component---cache-dev-404-page-js":["/component---cache-dev-404-page-js.js"],"component---src-pages-404-js":["/component---src-pages-404-js.js"],"component---src-pages-about-js":["/component---src-pages-about-js.js","/component---src-pages-about-js.f72585a3f4cb0feb1cf0.hot-update.js"],"component---src-pages-contact-js":["/component---src-pages-contact-js.js"],"component---src-pages-index-js":["/component---src-pages-index-js.js"],"component---src-pages-work-js":["/component---src-pages-work-js.js"]}
