+++ src/component/about/quick-bio/index.js
