+++ src/component/about/values-Missions/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "../layout/";
+import Layout from "../layout";
 
 function ValuesAndMissions() {
   return (
